Refactor MCP module tests into separate files and fix VM tests
