+++ orchestrator/src/mcp/transport.rs
@@ -261,310 +261,4 @@ impl Transport for StdioTransport {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::mcp::protocol::McpError;
-
-    // Helper to create a test request
-    fn create_test_request(id: u64, method: &str) -> McpRequest {
-        McpRequest::new(id, method, None)
-    }
-
-    // Helper to create a test response
-    fn create_test_response(id: u64, result: serde_json::Value) -> String {
-        format!(r#"{{"jsonrpc":"2.0","id":{},"result":{}}}"#, id, result)
-    }
-
-    #[cfg(unix)]
-    #[tokio::test]
-    async fn test_stdio_transport_send() {
-        // This test verifies serialization works, but doesn't actually spawn a process
-        // We'll test real spawning in integration tests
-        let request = create_test_request(1, "initialize");
-
-        // Verify the request can be serialized
-        let json = serde_json::to_string(&request).unwrap();
-        assert!(json.contains("\"jsonrpc\":\"2.0\""));
-        assert!(json.contains("\"method\":\"initialize\""));
-    }
-
-    #[tokio::test]
-    async fn test_stdio_transport_recv() {
-        // Test response deserialization
-        let response_json = create_test_response(1, serde_json::json!({"status": "ok"}));
-        let response: McpResponse = serde_json::from_str(&response_json).unwrap();
-
-        assert_eq!(response.id, 1);
-        assert!(response.is_success());
-        assert!(response.result.is_some());
-    }
-
-    #[cfg(unix)]
-    #[tokio::test]
-    async fn test_stdio_transport_recv_error() {
-        // Test error response deserialization
-        let error_json =
-            r#"{"jsonrpc":"2.0","id":1,"error":{"code":-32601,"message":"Method not found"}}"#;
-        let response: McpResponse = serde_json::from_str(error_json).unwrap();
-
-        assert_eq!(response.id, 1);
-        assert!(!response.is_success());
-        assert!(response.error.is_some());
-
-        let error = response.error.unwrap();
-        assert_eq!(error.code, -32601);
-        assert!(error.message.contains("Method not found"));
-    }
-
-    #[cfg(unix)]
-    #[tokio::test]
-    async fn test_stdio_transport_round_trip() {
-        // Test that we can serialize and deserialize correctly
-        let original_request = create_test_request(42, "tools/list");
-        let json = serde_json::to_string(&original_request).unwrap();
-        let deserialized_request: McpRequest = serde_json::from_str(&json).unwrap();
-
-        assert_eq!(original_request, deserialized_request);
-    }
-
-    #[test]
-    fn test_error_response_conversion() {
-        // Test that error responses convert correctly to Result
-        let error_response = McpResponse::err(1, McpError::method_not_found("test_method"));
-
-        assert!(!error_response.is_success());
-        let result = error_response.into_result();
-        assert!(result.is_err());
-
-        let error = result.unwrap_err();
-        assert_eq!(error.code, -32601);
-    }
-
-    #[cfg(unix)]
-    #[tokio::test]
-    async fn test_echo_server_mock() {
-        // This test demonstrates how the transport would work with a real process
-        // For now, we'll skip actual process spawning in unit tests
-        // Real integration tests will be in Task 1.5
-
-        // Create a mock echo server script (in /tmp)
-        let echo_script = r#"#!/bin/bash
-# Simple echo server that reads lines from stdin and writes them to stdout
-while IFS= read -r line; do
-    echo "$line"
-done
-"#;
-
-        let echo_path = "/tmp/mcp_echo_test.sh";
-        std::fs::write(echo_path, echo_script).unwrap();
-
-        #[cfg(unix)]
-        {
-            use tokio::process::Command;
-
-            // Make the script executable
-            Command::new("chmod")
-                .args(["+x", echo_path])
-                .output()
-                .await
-                .expect("Failed to make echo script executable");
-
-            // Spawn the echo server
-            let mut transport = StdioTransport::spawn(echo_path, &[])
-                .await
-                .expect("Failed to spawn echo server");
-
-            // Send a request
-            let request = create_test_request(1, "test");
-            transport
-                .send(&request)
-                .await
-                .expect("Failed to send request");
-
-            // Receive the echoed response
-            let response = transport.recv().await.expect("Failed to receive response");
-
-            // The echo server should echo back our JSON
-            assert_eq!(response.id, 1);
-
-            // Clean up
-            transport.kill().await.expect("Failed to kill echo server");
-
-            // Clean up the test file
-            let _ = std::fs::remove_file(echo_path);
-        }
-
-        #[cfg(not(unix))]
-        {
-            // Skip this test on non-Unix platforms
-            println!("Skipping echo server test on non-Unix platform");
-        }
-    }
-
-    #[cfg(not(windows))]
-    #[tokio::test]
-    async fn test_transport_kill_and_wait() {
-        // Test kill() and wait() methods
-        // We'll use a simple sleep command that we can kill
-
-        let echo_script = r#"#!/bin/bash
-# Sleep for a long time so we can kill it
-sleep 100
-"#;
-
-        let echo_path = "/tmp/mcp_kill_test.sh";
-        std::fs::write(echo_path, echo_script).unwrap();
-
-        {
-            use tokio::process::Command;
-
-            // Make the script executable
-            Command::new("chmod")
-                .args(["+x", echo_path])
-                .output()
-                .await
-                .expect("Failed to make script executable");
-
-            // Spawn the process
-            let mut transport = StdioTransport::spawn(echo_path, &[])
-                .await
-                .expect("Failed to spawn process");
-
-            // Kill the process
-            let result = transport.kill().await;
-            assert!(result.is_ok());
-
-            // Verify transport is disconnected
-            assert!(!transport.is_connected());
-
-            // Calling kill again should be ok (no-op)
-            let result2 = transport.kill().await;
-            assert!(result2.is_ok());
-
-            // Clean up
-            let _ = std::fs::remove_file(echo_path);
-        }
-    }
-
-    #[cfg(not(windows))]
-    #[tokio::test]
-    async fn test_transport_wait_without_kill() {
-        // Test wait() method without killing the process first
-        let echo_script = r#"#!/bin/bash
-# Exit immediately
-exit 42
-"#;
-
-        let echo_path = "/tmp/mcp_wait_test.sh";
-        std::fs::write(echo_path, echo_script).unwrap();
-
-        {
-            use tokio::process::Command;
-
-            // Make the script executable
-            Command::new("chmod")
-                .args(["+x", echo_path])
-                .output()
-                .await
-                .expect("Failed to make script executable");
-
-            // Spawn the process
-            let mut transport = StdioTransport::spawn(echo_path, &[])
-                .await
-                .expect("Failed to spawn process");
-
-            // Wait for the process to exit
-            let exit_code = transport.wait().await;
-            assert!(exit_code.is_ok());
-            assert_eq!(exit_code.unwrap(), Some(42));
-
-            // Verify transport is disconnected
-            assert!(!transport.is_connected());
-
-            // Clean up
-            let _ = std::fs::remove_file(echo_path);
-        }
-    }
-
-    #[test]
-    fn test_transport_trait_bounds() {
-        // Verify that StdioTransport implements the required trait bounds
-        fn assert_send_sync<T: Send + Sync>() {}
-        assert_send_sync::<StdioTransport>();
-    }
-
-    #[tokio::test]
-    async fn test_transport_send_when_disconnected() {
-        // This test verifies that send fails when transport is disconnected
-        // We can't easily test this with the real spawn, so we'll create a mock scenario
-        // by testing the error path logic
-        let result = serde_json::json!({});
-        let response_json = create_test_response(1, result);
-
-        // Verify the response can be deserialized
-        let _response: McpResponse = serde_json::from_str(&response_json).unwrap();
-    }
-
-    #[test]
-    fn test_transport_command() {
-        // Test the command() getter
-        let command_str = "test command with args";
-
-        // We can't easily test this without spawning, but we can verify
-        // the concept by checking that the command string format is correct
-        assert!(command_str.contains("test"));
-        assert!(command_str.contains("args"));
-    }
-
-    #[tokio::test]
-    async fn test_transport_recv_invalid_json() {
-        // Test that recv fails with invalid JSON
-        let invalid_json = r#"{"jsonrpc":"2.0","id":1,"invalid"#;
-        let result: std::result::Result<McpResponse, _> = serde_json::from_str(invalid_json);
-
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_transport_recv_missing_fields() {
-        // Test that recv fails with incomplete response
-        let incomplete = r#"{"jsonrpc":"2.0"}"#;
-        let result: std::result::Result<McpResponse, _> = serde_json::from_str(incomplete);
-
-        // This should fail because id is required
-        assert!(result.is_err());
-    }
-
-    #[cfg(not(windows))]
-    #[tokio::test]
-    async fn test_transport_command_getter() {
-        // Test that we can get the command string from a spawned transport
-        let echo_script = r#"#!/bin/bash
-echo "test"
-"#;
-
-        let echo_path = "/tmp/mcp_command_test.sh";
-        std::fs::write(echo_path, echo_script).unwrap();
-
-        {
-            use tokio::process::Command;
-
-            Command::new("chmod")
-                .args(["+x", echo_path])
-                .output()
-                .await
-                .expect("Failed to make script executable");
-
-            let transport = StdioTransport::spawn(echo_path, &[])
-                .await
-                .expect("Failed to spawn");
-
-            // Check that command() returns the command string
-            let cmd = transport.command();
-            assert!(cmd.contains(echo_path));
-
-            // Clean up
-            let _ = std::fs::remove_file(echo_path);
-        }
-    }
-}
+mod tests;
