+++ orchestrator/src/vm/mod.rs
@@ -30,15 +30,11 @@ use crate::vm::firewall::FirewallManager;
 use crate::vm::seccomp::{SeccompFilter, SeccompLevel};
 
 /// VM handle for managing lifecycle
-#[derive(Debug)]
 pub struct VmHandle {
     pub id: String,
-    #[allow(dead_code)] // Field is unused on Windows but required on Linux
     process: Arc<Mutex<Option<FirecrackerProcess>>>,
     pub spawn_time_ms: f64,
-    #[allow(dead_code)] // Field is unused on Windows but required on Linux
     config: VmConfig,
-    #[allow(dead_code)] // Field is unused on Windows but required on Linux
     firewall_manager: Option<FirewallManager>,
 }
 
@@ -221,179 +217,6 @@ pub async fn destroy_vm(handle: VmHandle) -> Result<()> {
 }
 
 #[cfg(test)]
-mod spawn_tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_vm_spawn_and_destroy() {
-        // This test requires actual Firecracker installation
-        // Skip in CI if not available
-        if !std::path::Path::new("/usr/local/bin/firecracker").exists() {
-            return;
-        }
-
-        // Ensure test assets exist
-        let _ = std::fs::create_dir_all("/tmp/ironclaw-fc-test");
-
-        let result = spawn_vm("test-task").await;
-
-        // If assets don't exist, we expect an error
-        if result.is_err() {
-            println!("Skipping test: Firecracker assets not available");
-            return;
-        }
-
-        let handle = result.unwrap();
-        assert_eq!(handle.id, "test-task");
-        assert!(handle.spawn_time_ms > 0.0);
-
-        destroy_vm(handle).await.unwrap();
-    }
-
-    #[test]
-    fn test_vm_id_format() {
-        let task_id = "task-123";
-        let expected_id = task_id.to_string();
-        assert_eq!(expected_id, "task-123");
-    }
-}
-
-/// Unit tests for VmHandle
-#[cfg(test)]
-mod vm_handle_tests {
-    use super::*;
-
-    #[test]
-    fn test_vm_handle_vsock_path_none() {
-        let mut config = VmConfig::default();
-        config.vsock_path = None; // Explicitly set to None
-        let handle = VmHandle {
-            id: "test-vm".to_string(),
-            process: std::sync::Arc::new(tokio::sync::Mutex::new(None)),
-            spawn_time_ms: 100.0,
-            config,
-            firewall_manager: None,
-        };
-
-        assert!(handle.vsock_path().is_none());
-    }
-
-    #[test]
-    fn test_vm_handle_vsock_path_some() {
-        let mut config = VmConfig::default();
-        config.vsock_path = Some("/tmp/test.sock".to_string());
-
-        let handle = VmHandle {
-            id: "test-vm".to_string(),
-            process: std::sync::Arc::new(tokio::sync::Mutex::new(None)),
-            spawn_time_ms: 100.0,
-            config,
-            firewall_manager: None,
-        };
-
-        assert_eq!(handle.vsock_path(), Some("/tmp/test.sock"));
-    }
-}
-
-/// Unit tests for verify_network_isolation
-#[cfg(test)]
-mod isolation_tests {
-    use super::*;
-
-    #[test]
-    fn test_verify_isolation_with_no_firewall_manager() {
-        let config = VmConfig::new("test-vm".to_string());
-        let handle = VmHandle {
-            id: "test-vm".to_string(),
-            process: std::sync::Arc::new(tokio::sync::Mutex::new(None)),
-            spawn_time_ms: 100.0,
-            config,
-            firewall_manager: None,
-        };
-
-        let result = verify_network_isolation(&handle);
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), false);
-    }
-
-    #[test]
-    fn test_verify_isolation_with_firewall_manager() {
-        let config = VmConfig::new("test-vm".to_string());
-        let firewall = FirewallManager::new("test-vm".to_string());
-        let handle = VmHandle {
-            id: "test-vm".to_string(),
-            process: std::sync::Arc::new(tokio::sync::Mutex::new(None)),
-            spawn_time_ms: 100.0,
-            config,
-            firewall_manager: Some(firewall),
-        };
-
-        let result = verify_network_isolation(&handle);
-        assert!(result.is_ok());
-    }
-}
-
-/// Unit tests for destroy_vm
-#[cfg(test)]
-mod destroy_tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_destroy_vm_with_no_process() {
-        let config = VmConfig::new("test-vm".to_string());
-        let handle = VmHandle {
-            id: "test-vm".to_string(),
-            process: std::sync::Arc::new(tokio::sync::Mutex::new(None)),
-            spawn_time_ms: 100.0,
-            config,
-            firewall_manager: None,
-        };
-
-        let result = destroy_vm(handle).await;
-        assert!(result.is_ok());
-    }
-}
-
-/// Unit tests for spawn_vm config logic
-#[cfg(test)]
-mod spawn_config_tests {
-    use super::*;
-
-    #[test]
-    fn test_spawn_vm_delegates_to_spawn_vm_with_config() {
-        // Test that spawn_vm creates a VmConfig and calls spawn_vm_with_config
-        // We can't actually test the async function here, but we can verify
-        // that VmConfig::new sets the expected values
-        let config = VmConfig::new("test-task".to_string());
-        assert_eq!(config.vm_id, "test-task");
-        assert!(config.vsock_path.is_some());
-    }
-
-    #[test]
-    fn test_vmconfig_seccomp_auto_enable_needed() {
-        // When seccomp_filter is None, Basic level should be auto-enabled
-        let config = VmConfig::default();
-        assert!(config.seccomp_filter.is_none());
-
-        // The logic in spawn_vm_with_config would add Basic seccomp
-        let should_add_seccomp = config.seccomp_filter.is_none();
-        assert!(should_add_seccomp);
-    }
-
-    #[test]
-    fn test_vmconfig_seccomp_already_set() {
-        // When seccomp_filter is Some, it should not be overridden
-        use seccomp::{SeccompFilter, SeccompLevel};
-        let config = VmConfig {
-            seccomp_filter: Some(SeccompFilter::new(SeccompLevel::Minimal)),
-            ..VmConfig::default()
-        };
-
-        // The logic in spawn_vm_with_config should keep the existing filter
-        let should_add_seccomp = config.seccomp_filter.is_none();
-        assert!(!should_add_seccomp);
-    }
-}
 
 /// Verify that a VM is properly network-isolated
 ///
